feat(review): 인증 구현

+++ backend/src/main/java/com/backend/domain/review/service/ReviewService.java
@@ -30,8 +30,7 @@ public class ReviewService {
 
     //리뷰 생성 메서드
     @Transactional
-    public ReviewResponseDto createReview(ReviewRequestDto reviewRequestDto){
-        long memberId = reviewRequestDto.memberId();
+    public ReviewResponseDto createReview(ReviewRequestDto reviewRequestDto, Long memberId) {
         long placeId = reviewRequestDto.placeId();
 
         Member member = getMemberEntity(memberId);
@@ -50,25 +49,48 @@ public class ReviewService {
 
     //리뷰 수정 메서드
     @Transactional
-    public void modifyReview(long memberId, int modifyRating){
-        Review review = reviewRepository.findByMemberId(memberId).orElseThrow(
-                () -> new BusinessException(ErrorCode.NOT_FOUND_REVIEW)
-        );
+    public void modifyReview(Long memberId, int modifyRating){
+//        Member member = getMemberEntity(memberId);
+//        Review review = reviewRepository.findByMemberId(member.getId()).orElseThrow(
+//                () -> new BusinessException(ErrorCode.NOT_FOUND_REVIEW)
+//        );
+        Review review = getReviewWithAuth(memberId);
         review.setRating(modifyRating);
         review.onUpdate();
     }
 
     //리뷰 삭제 메서드
     @Transactional
-    public void deleteReview(long reviewId){
+    public void deleteReview(Long memberId, long reviewId){
+//        Review review = getReviewEntity(reviewId);
+//        Member member = getMemberEntity(memberId);
+//        if(member.getId() != review.getMember().getId()){           //본인 검증? 이정도면 괜찮을지 걱정..
+//            throw new BusinessException(ErrorCode.ACCESS_DENIED);
+//        }
+        if(!validWithReviewId(memberId, reviewId)){
+            throw new BusinessException(ErrorCode.ACCESS_DENIED);
+        }
         Review review = getReviewEntity(reviewId);
         reviewRepository.delete(review);
     }
 
     //내가 작성한 리뷰 조회
-    public ReviewResponseDto getReview(long reviewId) {
-        Review review = getReviewEntity(reviewId);
-        return ReviewResponseDto.from(review);
+    public List<ReviewResponseDto> getMyReviews(Long memberId) {
+//        Member member = getMemberEntity(memberId);
+//        Review review = getReviewEntity(reviewId);
+//        Review review = reviewRepository.findByMemberId(member.getId()).orElseThrow(
+//                () -> new BusinessException(ErrorCode.NOT_FOUND_REVIEW)
+//        );
+        Review review = getReviewWithAuth(memberId);
+        Member member = getMemberEntity(memberId);
+        List<Review> myReviews = reviewRepository.findAllByMemberId(memberId);
+        if(myReviews.isEmpty()){
+            throw new BusinessException(ErrorCode.NOT_FOUND_REVIEW);
+        }
+        return myReviews.stream()
+                .map(ReviewResponseDto::from)
+                .toList();
+
     }
 
     //전체 리뷰 조회
@@ -80,14 +102,19 @@ public class ReviewService {
     }
 
     //여행지의 전체 리뷰 조회
-    public List<ReviewResponseDto> getReviewList(long placeId) {
+    public List<ReviewResponseDto> getReviewList(Long placeId) {
         return reviewRepository.findByPlaceId(placeId)
                 .stream()
                 .map(ReviewResponseDto::from)
                 .toList();
     }
 
-    public List<RecommendResponse> recommendByPlace(long placeId) {
+//    public List<RecommendResponse> recommendHotel() {
+//        Map<Long, Double> placeAverageRatings = new HashMap<>();
+//        List<Place> findAllPlaces = placeRepository.findByCategoryName("호텔");
+//    }
+
+    public List<RecommendResponse> recommendByPlace(Long placeId) {
         Map<Long, Double> placeAverageRatings = new HashMap<>(); //<placeId, averageRating> 으로 저장
 //        long placeSize = placeRepository.count();               //중간에 값이 삭제되고나면 id가 건더뛰게 되는 상황은 어떻게 처리? -> findAll()으로 변경
         List<Place> findAllPlaces = placeRepository.findAll();
@@ -111,22 +138,35 @@ public class ReviewService {
         return recommendList;
     }
 
-    public Review getReviewEntity(long reviewId){
+    public Review getReviewEntity(Long reviewId){
         return reviewRepository.findById(reviewId).orElseThrow(
                 () -> new BusinessException(ErrorCode.NOT_FOUND_REVIEW)
         );
     }
 
-    public Place getPlaceEntity(long placeId){
+    public Place getPlaceEntity(Long placeId){
         return placeRepository.findById(placeId).orElseThrow(
                 () -> new BusinessException(ErrorCode.NOT_FOUND_PLACE)
         );
     }
 
-    public Member getMemberEntity(long memberId){
+    public Member getMemberEntity(Long memberId){
         return memberRepository.findById(memberId).orElseThrow(
                 () -> new BusinessException(ErrorCode.MEMBER_NOT_FOUND)
         );
+    }
+
+    public Review getReviewWithAuth(Long memberId){
+        Member member = getMemberEntity(memberId);
+        return reviewRepository.findByMemberId(member.getId()).orElseThrow(
+                () -> new BusinessException(ErrorCode.NOT_FOUND_REVIEW)
+        );
+    }
+
+    public boolean validWithReviewId(Long memberId, Long reviewId){
+        Review review = getReviewEntity(reviewId);
+        Member member = getMemberEntity(memberId);
+        return review.getMember().getId().equals(member.getId());
     }
 
 }
